Allow DLL to be built before it's built
Previously, if you tried to build the dll before it was built, there was
an error in the non-dll file because it couldn't find it. This is
because the dll file uses the modules object from the non-dll object, so
that file needs to be compiled for it to work.

Instead, we check to see if the requiring filename is the dll or non-dll
file. If it is the non-dll file we don't include the DLLReferencePlugin,
otherwise we do and then everything builds fine.